refactor: shared domain type 분리 (#46)
