+++ src/features/shared/shared.hook.ts
@@ -1,7 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { useCallback, useMemo, useState } from 'react';
 
-import { getSharedPosts, type GetSharedPostsProps } from './shared.api';
+import { getSharedPosts } from './shared.api';
+import { type GetSharedPostsProps } from './shared.type';
 
 export const usePaging = ({
   maxPostPage,
