+++ src/features/shared/shared.type.ts
@@ -1,0 +1,11 @@
+import { type RentalType, type RoomType } from '@/shared/types';
+
+export interface GetSharedPostsFilter {
+  roomType?: RoomType[];
+  rentalType?: RentalType[];
+}
+
+export interface GetSharedPostsProps {
+  filter?: GetSharedPostsFilter;
+  search?: string;
+}
